+++ hydro_lang/Cargo.toml
@@ -24,6 +24,7 @@ viz = [
 ]
 deploy = [
     "build",
+    "deploy_integration",
     "dep:hydro_deploy",
     "dep:prettyplease",
     "dep:trybuild-internals-api",
@@ -32,6 +33,7 @@ deploy = [
     "dep:stageleft_tool",
     "dep:nameof",
 ]
+deploy_integration = ["dep:hydro_deploy_integration"]
 build = ["dep:dfir_lang", "dep:backtrace", "dep:ctor"]
 runtime_measure = ["dep:chrono", "dep:procfs"]
 runtime_support = ["dep:dfir_rs", "dfir_rs/deploy_integration"]
@@ -46,7 +48,7 @@ bytes = { version = "1.1.0", features = ["serde"] }
 bincode = "1.3.1"
 clap = { version = "4.0", features = ["derive"], optional = true }
 hydro_deploy = { path = "../hydro_deploy/core", version = "^0.14.0", optional = true }
-hydro_deploy_integration = { path = "../hydro_deploy/hydro_deploy_integration", version = "^0.14.0" }
+hydro_deploy_integration = { path = "../hydro_deploy/hydro_deploy_integration", version = "^0.14.0", optional = true }
 dfir_rs = { path = "../dfir_rs", version = "^0.14.0", default-features = false, optional = true }
 dfir_lang = { path = "../dfir_lang", version = "^0.14.0", optional = true }
 futures = "0.3.0"
