+++ hydro_lang/src/builder/built.rs
@@ -6,7 +6,7 @@ use dfir_lang::graph::{DfirGraph, eliminate_extra_unions_tees, partition_graph};
 
 use super::compiled::CompiledFlow;
 use super::deploy::{DeployFlow, DeployResult};
-use crate::deploy::{ClusterSpec, Deploy, ExternalSpec, IntoProcessSpec};
+use super::deploy_provider::{ClusterSpec, Deploy, ExternalSpec, IntoProcessSpec};
 #[cfg(feature = "viz")]
 use crate::graph::api::GraphApi;
 use crate::ir::{HydroRoot, emit};
